Atualizando e criando rotas

+++ src/controllers/PacienteController.js
@@ -1,3 +1,4 @@
+const mongoose = require('mongoose');
 const PacienteRepository = require("../repositories/PacienteRepository");
 
 class PacienteController {
@@ -6,7 +7,7 @@ class PacienteController {
         const pacientes = await PacienteRepository.findAll();
 
         if (!pacientes) {
-            return res.status(404).json({ error: "Pacientes não encontrados!" });
+            return res.status(404).json({ error: "Pacientes não encontrados!!!" });
         }
 
         res.json(pacientes);
@@ -15,10 +16,13 @@ class PacienteController {
     //Busca específico por ID -> criado por padrão, não será implementado no front-end
     async show(req, res) {
         const { id } = req.params;
-        const paciente = await PacienteRepository.findById(id);
+        if(!mongoose.Types.ObjectId.isValid(id)){ //valida para não entrar na rota errada
+            return res.status(400).json({ error: "ID inválido!!!" });
+        }
 
+        const paciente = await PacienteRepository.findById(id);
         if (!paciente) {
-            return res.status(404).json({ message: "Paciente com esse ID não encontrado!" });
+            return res.status(404).json({ message: "Paciente com esse ID não encontrado!!!" });
         }
 
         res.json(paciente);
@@ -30,7 +34,7 @@ class PacienteController {
         const paciente = await PacienteRepository.findByCpf(cpf);
 
         if (!paciente) {
-            return res.status(404).json({ message: "Paciente com esse CPF não encontrado!" });
+            return res.status(404).json({ message: "Paciente com esse CPF não encontrado!!!" });
         }
 
         res.json(paciente);
@@ -45,10 +49,10 @@ class PacienteController {
         }
 
         if(cpf) {
-            const pacienteCpf = await PacienteRepository.findByCPF(cpf);
+            const pacienteCpf = await PacienteRepository.findByCpf(cpf);
 
             if(pacienteCpf) {
-                return response.status(400).json({ error: "Esse CPF já está cadastrado!" });
+                return response.status(400).json({ error: "Esse CPF já está cadastrado!!!" });
             }
         }
 
@@ -62,18 +66,21 @@ class PacienteController {
     //Atualizar por ID -> criado por padrão, não será implementado no front-end
     async update(req, res) {
         const { id } = req.params;
-        const { nome, cpf, dataNascimento, endereco, telefone } = req.body;
+        if(!mongoose.Types.ObjectId.isValid(id)){
+            return res.status(400).json({ error: "ID inválido!!!" });
+        }
 
+        const { nome, cpf, dataNascimento, endereco, telefone } = req.body;
         const paciente = await PacienteRepository.findById(id);
         if(!paciente) {
             return res.status(404).json({ error: "Paciente não encontrado!!!"})
         }
 
         if(cpf) {
-            const pacienteCpf = await PacienteRepository.findByCPF(cpf);
+            const pacienteCpf = await PacienteRepository.findByCpf(cpf);
 
             if(pacienteCpf) {
-                return res.status(400).json({ error: "Esse CPF já está cadastrado em outro paciente!" });
+                return res.status(400).json({ error: "Esse CPF já está cadastrado em outro paciente!!!" });
             }
         }
 
@@ -93,22 +100,22 @@ class PacienteController {
         const { cpf } = req.params;
         const { nome, novoCpf, dataNascimento, endereco, telefone } = req.body;
 
-        const paciente = await PacienteRepository.findById(cpf);
+        const paciente = await PacienteRepository.findByCpf(cpf);
         if(!paciente) {
             return res.status(404).json({ error: "Paciente não encontrado!!!"})
         }
 
         if(novoCpf && novoCpf !== paciente.cpf) {
-            const pacienteCpf = await PacienteRepository.findByCPF(cpf);
+            const pacienteCpf = await PacienteRepository.findByCpf(novoCpf);
 
             if(pacienteCpf) {
-                return res.status(400).json({ error: "Esse CPF já está cadastrado em outro paciente!" });
+                return res.status(400).json({ error: "Esse CPF já está cadastrado em outro paciente!!!" });
             }
         }
 
         const pacienteAtualizado = await PacienteRepository.update(paciente._id, {
             nome: nome ?? paciente.nome,
-            cpf: cpf ?? paciente.cpf,
+            cpf: novoCpf ?? paciente.cpf,
             dataNascimento: dataNascimento ?? paciente.dataNascimento,
             endereco: endereco ?? paciente.endereco,
             telefone: telefone ?? paciente.telefone
@@ -120,6 +127,9 @@ class PacienteController {
     // Deletar por ID -> criado por padrão, não será implementado no front-end
     async destroy(req, res) {
         const { id } = req.params;
+        if(!mongoose.Types.ObjectId.isValid(id)){
+            return res.status(400).json({ error: "ID inválido!!!" });
+        }
 
         if(!id){
             return res.status(400).json({ error: "ID de Paciente inválido!!!"});
@@ -135,6 +145,11 @@ class PacienteController {
 
         if(!cpf){
             return res.status(400).json({ error: "CPF de Paciente inválido!!!"});
+        }
+
+        const paciente = await PacienteRepository.findByCpf(cpf)
+        if(!paciente){
+            return res.status(404).json({ error: "Paciente não encontrado!!!" });
         }
 
         await PacienteRepository.deleteByCpf(cpf);
